Make generateArgon2idHash freestanding
